+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Highlights are marked with a pancake 🥞
 - Remove in-memory store in `p2panda-store-next` and use SQLite in `p2panda-stream-next` [#1016](https://github.com/p2panda/p2panda/pull/1016)
 - Use `p2panda-store-next` SQLite stores in `p2panda-net` and `p2panda-sync` [#1022](https://github.com/p2panda/p2panda/pull/1022)
 - Use `Topic` everywhere [#1058](https://github.com/p2panda/p2panda/pull/1058)
+- More generic API for ingest & operation validation [#1050](https://github.com/p2panda/p2panda/pull/1050)
 
 ### Fixed